move things around a bit, test test contact case
