+++ temba/orgs/tests.py
@@ -15,10 +15,11 @@ from redis_cache import get_redis_connection
 from temba.campaigns.models import Campaign, CampaignEvent
 from temba.contacts.models import Contact, ContactGroup, TEL_SCHEME, TWITTER_SCHEME
 from temba.orgs.models import Org, OrgCache, OrgEvent, OrgFolder, TopUp, Invitation, DAYFIRST, MONTHFIRST
+from temba.orgs.models import UNREAD_FLOW_MSGS, UNREAD_INBOX_MSGS
 from temba.channels.models import Channel, RECEIVE, SEND, TWILIO, TWITTER, PLIVO_AUTH_ID, PLIVO_AUTH_TOKEN
 from temba.flows.models import Flow, ActionSet
 from temba.msgs.models import Broadcast, Call, Label, Msg, Schedule, CALL_IN, INCOMING
-from temba.tests import TembaTest, MockResponse, MockTwilioClient, MockRequestValidator
+from temba.tests import TembaTest, MockResponse, MockTwilioClient, MockRequestValidator, FlowFileTest
 from temba.triggers.models import Trigger
 
 
@@ -1405,3 +1406,68 @@ class BulkExportTest(TembaTest):
         self.assertEquals(60, confirm_appointment.expires_after_minutes)
 
 
+class UnreadCountTest(FlowFileTest):
+
+    def test_unread_count_test(self):
+        flow = self.get_flow('favorites')
+
+        # create a trigger for 'favs'
+        Trigger.objects.create(org=self.org, flow=flow, keyword='favs', created_by=self.admin, modified_by=self.admin)
+
+        # start our flow by firing an incoming message
+        contact = self.create_contact('Anakin Skywalker', '+12067791212')
+        msg = self.create_msg(contact=contact, text="favs")
+
+        # process it
+        Msg.process_message(msg)
+
+        # at this point our flow should have started.. go to our trigger list page to see if our context is correct
+        self.login(self.admin)
+        trigger_list = reverse('triggers.trigger_list')
+        response = self.client.get(trigger_list)
+
+        self.assertEquals(0, response.context['msgs_unread_count'])
+        self.assertEquals(1, response.context['flows_unread_count'])
+
+        # answer another question in the flow
+        msg = self.create_msg(contact=contact, text="red")
+        Msg.process_message(msg)
+
+        response = self.client.get(trigger_list)
+        self.assertEquals(0, response.context['msgs_unread_count'])
+        self.assertEquals(2, response.context['flows_unread_count'])
+
+        # finish the flow and send a message outside it
+        msg = self.create_msg(contact=contact, text="primus")
+        Msg.process_message(msg)
+
+        msg = self.create_msg(contact=contact, text="nic")
+        Msg.process_message(msg)
+
+        msg = self.create_msg(contact=contact, text="Hello?")
+        Msg.process_message(msg)
+
+        response = self.client.get(trigger_list)
+        self.assertEquals(4, response.context['flows_unread_count'])
+        self.assertEquals(1, response.context['msgs_unread_count'])
+
+        # visit the msg pane
+        response = self.client.get(reverse('msgs.msg_inbox'))
+        self.assertEquals(4, response.context['flows_unread_count'])
+        self.assertEquals(0, response.context['msgs_unread_count'])
+
+        # now the flow list pane
+        response = self.client.get(reverse('flows.flow_list'))
+        self.assertEquals(0, response.context['flows_unread_count'])
+        self.assertEquals(0, response.context['msgs_unread_count'])
+
+        # make sure a test contact doesn't update our counts
+        contact.is_test = True
+        contact.save()
+
+        msg = self.create_msg(contact=contact, text="favs")
+        Msg.process_message(msg)
+
+        # assert our counts weren't updated
+        self.assertEquals(0, self.org.get_unread_msg_count(UNREAD_INBOX_MSGS))
+        self.assertEquals(0, self.org.get_unread_msg_count(UNREAD_FLOW_MSGS))
